feat: add text-based sell task creation
